+++ CodeStarterCamp_Week5/main.swift
@@ -8,68 +8,12 @@
 
 import Foundation
 
-class Person {
-    var name: String
-    var height: Int
-    
-    init(name: String, height: Int) {
-        self.name = name
-        self.height = height
-    }
-}
-
-enum Level {
-    case A, B, C
-}
-
-protocol Talent {
-    var singing: Level { get }
-    var dancing: Level { get }
-    var acting: Level { get }
-}
-
-protocol BadPersonality {
-    var frequancyOfCursing: Level { get }
-}
-
-class TalentedPerson: Person, Talent {
-    var singing: Level
-    var dancing: Level
-    var acting: Level
-    
-    init(name: String, height: Int, singing: Level, dancing: Level, acting: Level) {
-        self.singing = singing
-        self.dancing = dancing
-        self.acting = acting
-        
-        super.init(name: name, height: height)
-    }
-}
-
-class TalentedPersonWithBadPersonality: Person, Talent, BadPersonality {
-    var singing: Level
-    var dancing: Level
-    var acting: Level
-    var frequancyOfCursing: Level
-    
-    init(name: String, height: Int, singing: Level, dancing: Level, acting: Level, cursing: Level) {
-        self.singing = singing
-        self.dancing = dancing
-        self.acting = acting
-        self.frequancyOfCursing = cursing
-        
-        super.init(name: name, height: height)
-    }
-}
-
-
 let yagom = TalentedPerson(name: "yagom", height: 100, singing: .B, dancing: .A, acting: .C)
 let noroo = Person(name: "noroo", height: 1000)
 let summer = TalentedPerson(name: "summer", height: 900, singing: .B, dancing: .B, acting: .B)
 let coda = TalentedPerson(name: "coda", height: 200, singing: .A, dancing: .C, acting: .C)
 let odong = TalentedPersonWithBadPersonality(name: "odong", height: 400, singing: .A, dancing: .A, acting: .A, cursing: .A)
 
-var whalesEntertainment = AuditionManager(totalApplicantsList: [yagom, noroo, summer, coda, odong])
-whalesEntertainment.cast(applicants: whalesEntertainment.totalApplicantsList)
-whalesEntertainment.announcePassedApplicants(passedList: whalesEntertainment.passedApplicantsList)
-
+var auditionManager = AuditionManager()
+auditionManager.cast(applicants: auditionManager.totalApplicantsList)
+auditionManager.announcePassedApplicants()
